+++ aruw-mcb-project/src/aruwsrc/robot/sentry/sentry_response_transmitter.cpp
@@ -41,7 +41,7 @@ bool SentryResponseTransmitter::send()
 
         if ((queuedMessageType & (1 << static_cast<uint8_t>(lastSentMessage))) != 0)
         {
-            *reinterpret_cast<uint16_t*>(this->robotToRobotMessage.dataAndCRC16) = static_cast<uint8_t>(lastSentMessage);
+            *reinterpret_cast<uint16_t*>(this->robotToRobotMessage.dataAndCRC16) = static_cast<uint16_t>(lastSentMessage);
 
             PT_CALL(refSerialTransmitter.sendRobotToRobotMsg(
                 &robotToRobotMessage,
